Introduced sub implementation folder for point methods
